refactor(route paths): move helper utils into lib folder
