+++ src/app/routes/Router.tsx
@@ -19,13 +19,17 @@ import {
     SchemaSection,
     OVERVIEW_SECTIONS,
 } from '../../constants'
-import { isValidUid, isModuleNotFoundError } from '../../lib'
+import {
+    getSectionPath,
+    isModuleNotFoundError,
+    isValidUid,
+    routePaths,
+} from '../../lib'
 import { OverviewSection } from '../../types'
 import { Layout, Breadcrumbs, BreadcrumbItem } from '../layout'
 import { CheckAuthorityForSection } from './CheckAuthorityForSection'
 import { DefaultErrorRoute } from './DefaultErrorRoute'
 import { LegacyAppRedirect } from './LegacyAppRedirect'
-import { getSectionPath, routePaths } from './routePaths'
 
 // This loads all the overview routes in the same chunk since they resolve to the same promise
 // see https://reactrouter.com/en/main/route/lazy#multiple-routes-in-a-single-file
