+++ src/lib/routeUtils/index.ts
@@ -1,2 +1,8 @@
 export * from './CustomQueryParam'
+export {
+    routePaths,
+    getOverviewPath,
+    getSectionPath,
+    getSectionNewPath,
+} from './routePaths'
 export * from './useSectionHandle'
